Solve int overflow bug

+++ ConsoleApp1/Program.cs
@@ -1,4 +1,4 @@
 ﻿int x = 2100000000;
 int y = 2100000000;
-int z = x + y;
+long z = (long)x + y;
 Console.Write(z);
